New Layout Code: I've added a 2D array "layout", which is basically trying to be the magic file (the picture of the layout we've been referencing). A CSV is generated from it, which when pasted into the shared spreadsheet, turns colors and stuff so you can see what's going on. Right now only cells are in the layout, but the routes will get added in due time and then you'll be able to see everything and how it's connected. The layout codes (0-9) refer to (0) empty space, (1-4) rotations of cells, (5,6) rotations of pass-thru cells, (7,8) metal 1 and metal 2, (9) vias.
In main.cpp I've carved out a big block where I've started to test the layout formation by just putting your mainPartition in a single row. We will want to modify this code (and probably move out of main.cpp) so that it places the cells in a 2:1 rectangle before I begin routing.
In addition to the 2D layout vector (since it only stores cells' rotation info) I have a vector cellData. This is the real important one. It's a vector of the "cell" structure I created in structures.h, which holds info about each of the cells in our design. It keeps track of the lower left corner coordinates of the cell so that we can map it to the layout, rotation, and number of nets. Trying to keep cellData vector organized so that the cellData[i] is cell i. We can use the cellData vector to store dummies and pass-thru cells as well, but we'll just tack them on the end of the vector after all the actual cells.